+++ src/defaults/config.ts
@@ -1,0 +1,20 @@
+import { LIGHT_THEME } from '../styles/light.theme';
+import { AppLangs, AppTheme, MainContentTab } from '../types';
+
+export const DEFAULTS = {
+    APP_THEME: AppTheme.LIGHT,
+    THEME: LIGHT_THEME,
+    LANG: AppLangs.EN,
+    LANGS_AMOUNT: 2,
+
+    EMPTY_ROLL_RESULT: null,
+
+    DICE_MIN: 1,
+    DICE_NUMBER: 1,
+    MOD: 0,
+
+    SHOW_INPUT_MOD_ON_ENTER: false,
+    MAIN_CONTENT: MainContentTab.FATE, //TODO: for test only!
+
+    REPO_URL: 'https://github.com/NdV66/roll-my-dices',
+};
