+++ src/defaults/cookies.ts
@@ -1,0 +1,2 @@
+export const COOKIE_THEME_KEY = 'themeKey';
+export const COOKIE_LANG_KEY = 'langKey';
