+++ src/defaults/errors.ts
@@ -1,0 +1,3 @@
+export enum ERROR_CODES {
+    'NO_DICE_FOUND' = 'No dice sign found',
+}
