+++ src/defaults/fate.ts
@@ -1,0 +1,20 @@
+export const FATE = {
+    PLUS: 1,
+    MINUS: -1,
+    NEUTRAL: 0,
+
+    DICE_NUMBERS: 4,
+
+    TRANSLATE_FOR_MINUS: [1, 2],
+    TRANSLATE_FOR_NEUTRAL: [3, 4],
+    TRANSLATE_FOR_PLUS: [5, 6],
+
+    MIN: 1,
+    MAX: 6,
+
+    SIGNS: {
+        PLUS: '+',
+        MINUS: '-',
+        NEUTRAL: '',
+    },
+};
