+++ src/defaults/index.ts
@@ -1,0 +1,5 @@
+export * from './errors';
+export * from './cookies';
+export * from './dices';
+export * from './fate';
+export * from './config';
